dq_process_monitoring/requirements.txt
argcomplete==1.12.2
boto==2.49.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crcmod==1.7
croniter==1.0.8
cryptography==3.4.6
fasteners==0.16
fsspec==0.8.7
future==0.18.2
gcloud==0.18.3
gcs-oauth2-boto-plugin==2.7
gcsfs==0.7.2
google-api-core==1.26.1
google-apitools==0.5.31
google-auth==1.27.1
google-cloud-bigquery==2.11.0
google-cloud-bigquery-storage==2.3.0
google-cloud-core==1.6.0
google-cloud-logging==2.2.0
google-cloud-pubsub==2.3.0
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-reauth==0.1.1
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.36.1
gsutil==4.60
httplib2==0.19.0
idna==2.10
libcst==0.3.17
mock==2.0.0
monotonic==1.5
mypy-extensions==0.4.3
natsort==7.1.1
numpy==1.20.1
oauth2client==4.1.3
packaging==20.9
pandas==1.2.3
pbr==5.5.1
proto-plus==1.17.0
protobuf==3.15.6
psycopg2-binary==2.8.6
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pyu2f==0.1.5
PyYAML==5.4.1
requests==2.25.1
retry-decorator==1.1.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.23
tqdm==4.59.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3

